use own lockfile package and do not remove file
